+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 Changelog
 =========
-## 0.6.0 (2017-07-xx)
+## 0.6.0 (2017-12-01)
 * BC BREAK: Fully replaced Buzz library with usage of HTTPlug & Guzzle 6,
 * BC BREAK: `hwi.http_client` config options are remove. HTTP configuration must rely on the HTTPlug client,
 * BC BREAK: Template engine other than Twig are no longer supported,
 * BC BREAK: Option `hwi_oauth.templating_engine` was removed,
+* Added: Symfony 4 support,
 * Added: `php-http/httplug-bundle` support, to auto-provide needed HTTPlug services and get full Symfony integration,
 * Added: `hwi.http.client` and `hwi.http.message_factory` config keys to provide your own HTTPlug services,
 * Added: `HWIOAuthEvents`,